+++ HISTORY.md
@@ -5,6 +5,7 @@
 * Replaced Maven with Gradle for building
 * Added support for recurring cycle numbers
 * Added support for default_email_sender and print_version in Settings
+* Added support for client contacts
 
 ## 1.0.4 (2013-04-13)